+++ account_advanced_reconcile_transaction_by_purchase_line/README.rst
@@ -1,0 +1,67 @@
+.. image:: https://img.shields.io/badge/license-AGPLv3-blue.svg
+   :target: https://www.gnu.org/licenses/agpl.html
+   :alt: License: AGPL-3
+
+=======================================================
+Account Advanced Reconcile Transaction by Purchase Line
+=======================================================
+
+This module will allow a user to reconcile debits and credits of an Account
+using the PO Line, Product and Partner as key fields. This type of
+reconciliation is to be used in the context of the Perpetual Inventory
+accounting system, with the accrual account 'Goods Received Not Invoiced'.
+
+
+Usage
+=====
+
+* Go to 'Invoicing / Periodic Processing / Reconciliation / Easy Automatic
+Reconciliation'.
+
+* Create a new reconciliation profile, and select a new configuration entry
+with type 'Advanced. GR/IR Key as partner, product, purchase order line.'.
+
+.. image:: https://odoo-community.org/website/image/ir.attachment/5784_f2813bd/datas
+   :alt: Try me on Runbot
+   :target: https://runbot.odoo-community.org/runbot/98/8.0
+
+Bug Tracker
+===========
+
+Bugs are tracked on `GitHub Issues
+<https://github.com/OCA/98/issues>`_. In case of trouble, please
+check there if your issue has already been reported. If you spotted it first,
+help us smashing it by providing a detailed and welcomed `feedback
+<https://github.com/OCA/
+98/issues/new?body=module:%20
+account_advanced_reconcile_transaction_by_purchase_line%0Aversion:%20
+8.0%0A%0A**Steps%20to%20reproduce**%0A-%20..
+.%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
+
+Credits
+=======
+
+Images
+------
+
+* Odoo Community Association: `Icon <https://github.com/OCA/maintainer-tools/blob/master/template/module/static/description/icon.svg>`_.
+
+Contributors
+------------
+
+* Jordi Ballester Alomar <jordi.ballester@eficent.com>
+
+Maintainer
+----------
+
+.. image:: https://odoo-community.org/logo.png
+   :alt: Odoo Community Association
+   :target: http://odoo-community.org
+
+This module is maintained by the OCA.
+
+OCA, or the Odoo Community Association, is a nonprofit organization whose
+mission is to support the collaborative development of Odoo features and
+promote its widespread use.
+
+To contribute to this module, please visit http://odoo-community.org.
