--- account_advanced_reconcile_transaction_by_purchase_line/base_advanced_reconciliation.py
@@ -1,37 +1,0 @@
-# -*- coding: utf-8 -*-
-##############################################################################
-#
-#    Author: Romain Deheele
-#    Copyright 2013 Camptocamp SA
-#
-#    This program is free software: you can redistribute it and/or modify
-#    it under the terms of the GNU Affero General Public License as
-#    published by the Free Software Foundation, either version 3 of the
-#    License, or (at your option) any later version.
-#
-#    This program is distributed in the hope that it will be useful,
-#    but WITHOUT ANY WARRANTY; without even the implied warranty of
-#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#    GNU Affero General Public License for more details.
-#
-#    You should have received a copy of the GNU Affero General Public License
-#    along with this program.  If not, see <http://www.gnu.org/licenses/>.
-#
-##############################################################################
-
-from openerp import models, api
-
-
-class EasyReconcileAdvanced(models.AbstractModel):
-
-    _inherit = 'easy.reconcile.advanced'
-
-    @api.model
-    def _base_columns(self):
-        """ Mandatory columns for move lines queries
-        An extra column aliased as ``key`` should be defined
-        in each query."""
-        aml_cols = super(EasyReconcileAdvanced, self)._base_columns()
-        aml_cols.append('account_move_line.purchase_line_id')
-        aml_cols.append('account_move_line.product_id')
-        return aml_cols
