--- account_advanced_reconcile_transaction_by_purchase_line/easy_reconcile.py
@@ -1,36 +1,0 @@
-# -*- coding: utf-8 -*-
-##############################################################################
-#
-#    Author: Romain Deheele
-#    Copyright 2013 Camptocamp SA
-#
-#    This program is free software: you can redistribute it and/or modify
-#    it under the terms of the GNU Affero General Public License as
-#    published by the Free Software Foundation, either version 3 of the
-#    License, or (at your option) any later version.
-#
-#    This program is distributed in the hope that it will be useful,
-#    but WITHOUT ANY WARRANTY; without even the implied warranty of
-#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#    GNU Affero General Public License for more details.
-#
-#    You should have received a copy of the GNU Affero General Public License
-#    along with this program.  If not, see <http://www.gnu.org/licenses/>.
-#
-##############################################################################
-from openerp import models, api
-
-
-class AccountEasyReconcileMethod(models.Model):
-
-    _inherit = 'account.easy.reconcile.method'
-
-    @api.model
-    def _get_all_rec_method(self):
-        methods = super(AccountEasyReconcileMethod, self).\
-            _get_all_rec_method()
-        methods += [
-            ('easy.reconcile.advanced.by.purchase.line',
-             'Advanced. GR/IR Key as partner, product, purchase order line.'),
-        ]
-        return methods
