+++ account_advanced_reconcile_transaction_by_purchase_line/wizards/__init__.py
@@ -2,5 +2,4 @@
 # © 2015 Eficent Business and IT Consulting Services S.L. (www.eficent.com)
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html).
 
-from . import models
-from . import wizards
+from . import advanced_reconciliation
